pyproject.toml
[tool.poetry]
name = "fisco-tech"
version = "0.1.0"
description = ""
authors = ["Gustavo Monteiro <monteiro@tins.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
django-rest-framework = "^0.1.0"
lxml = "^5.1.0"
xmlschema = "^3.0.1"
drf-writable-nested = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
